Move standalone jar to new module, move frontend/backend routing code to standalone module
