--- employee-rostering-frontend/assembly/sources.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<assembly xmlns="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.2" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-          xsi:schemaLocation="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.2 http://maven.apache.org/xsd/assembly-1.1.2.xsd">
-
-  <id>sources</id>
-  <formats>
-    <format>jar</format>
-  </formats>
-
-  <includeBaseDirectory>false</includeBaseDirectory>
-  
-  <files>
-    <file>
-      <source>package.json</source>
-    </file>
-    <file>
-      <source>package-lock.json</source>
-    </file>
-    <file>
-      <source>tsconfig.json</source>
-    </file>
-  </files>
-  <fileSets>
-    <fileSet>
-      <directory>src</directory>
-      <outputDirectory>src</outputDirectory>
-    </fileSet>
-    <fileSet>
-      <directory>translations</directory>
-      <outputDirectory>translations</outputDirectory>
-    </fileSet>
-    <fileSet>
-      <directory>public</directory>
-      <outputDirectory>public</outputDirectory>
-    </fileSet>
-  </fileSets>
-</assembly>
