+++ employee-rostering-standalone/.gitignore
@@ -1,0 +1,11 @@
+/target
+/local
+
+# Eclipse, Netbeans and IntelliJ files
+/.*
+!.gitignore
+!.gitattributes
+/nbproject
+/*.ipr
+/*.iws
+/*.iml
